+++ packages/plugin-rax-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 6.4.6
+
+- Feat: merge postcssOptions support
+
+## 6.4.5
+
+- Fix: tabbar state update
+
+## 6.4.4
+
+- Fix: ssr inline style
+
+## 6.4.3
+
+- Feat: add config option `forceEnableCSS` support
+
 ## 6.4.2
 
 - Fix: tabBar height in iphoneX